kis refactoring
FormTestsBase absztrakt osztály létrehozva. A többi TestFixture osztályt erre építettem.
