+++ ClientAppUnitTests/GlobalUsings.cs
@@ -1,1 +1,2 @@
 global using NUnit.Framework;
+global using RaktarKeszletDasHaus;
